feat(core): clean up remaining usages of angular devkit
