+++ packages/create-nx-workspace/.eslintrc.json
@@ -1,5 +1,13 @@
 {
   "extends": "../../.eslintrc",
-  "rules": {},
+  "rules": {
+    "no-restricted-imports": [
+      "error",
+      "@nrwl/workspace",
+      "@angular-devkit/core",
+      "@angular-devkit/architect",
+      "@angular-devkit/schematics"
+    ]
+  },
   "ignorePatterns": ["!**/*"]
 }
