+++ packages/workspace/.eslintrc.json
@@ -1,5 +1,20 @@
 {
   "extends": "../../.eslintrc",
   "rules": {},
+  "overrides": [
+    {
+      "files": ["**/*.ts"],
+      "excludedFiles": ["./src/migrations/**", "./src/utils/**"],
+      "rules": {
+        "no-restricted-imports": [
+          "error",
+          "@nrwl/workspace",
+          "@angular-devkit/core",
+          "@angular-devkit/architect",
+          "@angular-devkit/schematics"
+        ]
+      }
+    }
+  ],
   "ignorePatterns": ["!**/*"]
 }
