+++ packages/workspace/src/generators/init/files/root/tools/tsconfig.tools.json
@@ -5,7 +5,8 @@
     "rootDir": ".",
     "module": "commonjs",
     "target": "es5",
-    "types": ["node"]
+    "types": ["node"],
+    "importHelpers": false
   },
   "include": ["**/*.ts"]
 }
