+++ packages/workspace/src/generators/workspace-generator/workspace-generator.ts
@@ -5,19 +5,32 @@ import {
   generateFiles,
   names,
   joinPathFragments,
+  addDependenciesToPackageJson,
 } from '@nrwl/devkit';
+import { nxVersion } from '../../utils/versions';
 
 export default async function (host: Tree, schema: Schema) {
   const options = normalizeOptions(schema);
+
   generateFiles(
     host,
     joinPathFragments(__dirname, 'files'),
     joinPathFragments('tools/generators', schema.name),
     options
   );
+
+  const installTask = addDependenciesToPackageJson(
+    host,
+    {},
+    {
+      '@nrwl/devkit': nxVersion,
+    }
+  );
+
   if (!schema.skipFormat) {
     await formatFiles(host);
   }
+  return installTask;
 }
 
 function normalizeOptions(options: Schema): any {
