+++ packages/workspace/src/generators/workspace/files/tools/tsconfig.tools.json
@@ -5,7 +5,8 @@
     "rootDir": ".",
     "module": "commonjs",
     "target": "es5",
-    "types": ["node"]
+    "types": ["node"],
+    "importHelpers": false
   },
   "include": ["**/*.ts"]
 }
