+++ packages/workspace/src/utilities/typescript/find-nodes.ts
@@ -1,0 +1,36 @@
+import * as ts from 'typescript';
+
+export function findNodes(
+  node: ts.Node,
+  kind: ts.SyntaxKind | ts.SyntaxKind[],
+  max = Infinity
+): ts.Node[] {
+  if (!node || max == 0) {
+    return [];
+  }
+
+  const arr: ts.Node[] = [];
+  const hasMatch = Array.isArray(kind)
+    ? kind.includes(node.kind)
+    : node.kind === kind;
+  if (hasMatch) {
+    arr.push(node);
+    max--;
+  }
+  if (max > 0) {
+    for (const child of node.getChildren()) {
+      findNodes(child, kind, max).forEach((node) => {
+        if (max > 0) {
+          arr.push(node);
+        }
+        max--;
+      });
+
+      if (max <= 0) {
+        break;
+      }
+    }
+  }
+
+  return arr;
+}
